move dedup from ascwds to clean cqc ind
