+++ jobs/clean_ascwds_workplace_data.py
@@ -66,17 +66,6 @@ def main(source: str, destination: str):
         AscwdsScaleVariableLimits.worker_records_lower_limit,
     )
 
-    ascwds_workplace_df = create_column_with_repeated_values_removed(
-        ascwds_workplace_df,
-        AWPClean.total_staff_bounded,
-        AWPClean.total_staff_deduplicated,
-    )
-    ascwds_workplace_df = create_column_with_repeated_values_removed(
-        ascwds_workplace_df,
-        AWPClean.worker_records_bounded,
-        AWPClean.worker_records_deduplicated,
-    )
-
     print(f"Exporting as parquet to {destination}")
     utils.write_to_parquet(
         ascwds_workplace_df,
@@ -174,54 +163,6 @@ def calculate_latest_update_to_workplace_location(df: DataFrame, comparison_date
     df_with_latest_update = df_with_latest_update.drop("latest_org_mupddate")
 
     return df_with_latest_update
-
-
-def create_column_with_repeated_values_removed(
-    df: DataFrame,
-    column_to_clean: str,
-    new_column_name: str = None,
-) -> DataFrame:
-    """
-    ASCWDS repeats data until it is changed. This function creates a new column which converts repeated values to nulls,
-    so we only see newly submitted values once.
-
-    For each workplace, this function iterates over the dataframe in date order and compares the current column value to the
-    previously submitted value. If the value differs from the previously submitted value then enter that value into the new column.
-    Otherwise null the value in the new column as it is a previously submitted value which has been repeated.
-
-    Args:
-        df: The dataframe to use
-        column_to_clean: The name of the column to convert
-        new_column_name: (optional) If not provided, "_deduplicated" will be appended onto the original column name
-
-    Returns:
-        A DataFrame with an addional column with repeated values changed to nulls.
-    """
-    PREVIOUS_VALUE: str = "previous_value"
-
-    if new_column_name is None:
-        new_column_name = column_to_clean + "_deduplicated"
-
-    w = Window.partitionBy(AWPClean.establishment_id).orderBy(
-        AWPClean.ascwds_workplace_import_date
-    )
-
-    df_with_previously_submitted_value = df.withColumn(
-        PREVIOUS_VALUE, F.lag(column_to_clean).over(w)
-    )
-
-    df_without_repeated_values = df_with_previously_submitted_value.withColumn(
-        new_column_name,
-        F.when(
-            (F.col(PREVIOUS_VALUE).isNull())
-            | (F.col(column_to_clean) != F.col(PREVIOUS_VALUE)),
-            F.col(column_to_clean),
-        ).otherwise(None),
-    )
-
-    df_without_repeated_values = df_without_repeated_values.drop(PREVIOUS_VALUE)
-
-    return df_without_repeated_values
 
 
 def purge_outdated_workplaces(
