+++ tests/test_file_schemas.py
@@ -279,23 +279,6 @@ class ASCWDSWorkplaceSchemas:
         ]
     )
 
-    repeated_value_schema = StructType(
-        [
-            StructField(AWPClean.establishment_id, StringType(), True),
-            StructField("integer_column", IntegerType(), True),
-            StructField(AWPClean.ascwds_workplace_import_date, DateType(), True),
-        ]
-    )
-
-    expected_without_repeated_values_schema = StructType(
-        [
-            StructField(AWPClean.establishment_id, StringType(), True),
-            StructField("integer_column", IntegerType(), True),
-            StructField(AWPClean.ascwds_workplace_import_date, DateType(), True),
-            StructField("integer_column_deduplicated", IntegerType(), True),
-        ]
-    )
-
 
 @dataclass
 class ONSData:
@@ -802,6 +785,23 @@ class CleanIndCQCData:
             StructField(IndCQC.worker_records_bounded, IntegerType(), True),
             StructField(IndCQC.ascwds_filled_posts, DoubleType(), True),
             StructField(IndCQC.ascwds_filled_posts_source, StringType(), True),
+        ]
+    )
+
+    repeated_value_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), True),
+            StructField("integer_column", IntegerType(), True),
+            StructField(IndCQC.cqc_location_import_date, DateType(), True),
+        ]
+    )
+
+    expected_without_repeated_values_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), True),
+            StructField("integer_column", IntegerType(), True),
+            StructField(IndCQC.cqc_location_import_date, DateType(), True),
+            StructField("integer_column_deduplicated", IntegerType(), True),
         ]
     )
