+++ tests/unit/test_clean_ascwds_workplace_data.py
@@ -223,80 +223,6 @@ class PurgeOutdatedWorkplacesColumn(AddPurgeOutdatedWorkplacesColumnTests):
         self.assertEqual(purge_data_list, expected_list)
 
 
-class AddColumnWithRepeatedValuesRemovedTests(IngestASCWDSWorkerDatasetTests):
-    def setUp(self):
-        super().setUp()
-        self.test_purge_outdated_df = self.spark.createDataFrame(
-            Data.repeated_value_rows, Schemas.repeated_value_schema
-        )
-        self.expected_df_without_repeated_values_df = self.spark.createDataFrame(
-            Data.expected_without_repeated_values_rows,
-            Schemas.expected_without_repeated_values_schema,
-        )
-        self.returned_df = job.create_column_with_repeated_values_removed(
-            self.test_purge_outdated_df,
-            column_to_clean="integer_column",
-        )
-        self.OUTPUT_COLUMN = "integer_column_deduplicated"
-
-        self.returned_data = self.returned_df.sort(
-            AWPClean.establishment_id, AWPClean.ascwds_workplace_import_date
-        ).collect()
-        self.expected_data = self.expected_df_without_repeated_values_df.sort(
-            AWPClean.establishment_id, AWPClean.ascwds_workplace_import_date
-        ).collect()
-
-    def test_first_submitted_value_is_included_in_new_column(self):
-        self.assertEqual(
-            self.returned_data[0][self.OUTPUT_COLUMN],
-            self.expected_data[0][self.OUTPUT_COLUMN],
-        )
-        self.assertEqual(
-            self.returned_data[4][self.OUTPUT_COLUMN],
-            self.expected_data[4][self.OUTPUT_COLUMN],
-        )
-
-    def test_submitted_value_is_included_if_it_wasnt_repeated(self):
-        self.assertEqual(
-            self.returned_data[1][self.OUTPUT_COLUMN],
-            self.expected_data[1][self.OUTPUT_COLUMN],
-        )
-        self.assertEqual(
-            self.returned_data[5][self.OUTPUT_COLUMN],
-            self.expected_data[5][self.OUTPUT_COLUMN],
-        )
-
-    def test_repeated_value_entered_as_null_value(self):
-        self.assertEqual(
-            self.returned_data[2][self.OUTPUT_COLUMN],
-            self.expected_data[2][self.OUTPUT_COLUMN],
-        )
-        self.assertEqual(
-            self.returned_data[7][self.OUTPUT_COLUMN],
-            self.expected_data[7][self.OUTPUT_COLUMN],
-        )
-
-    def test_value_which_has_appeared_before_but_isnt_a_repeat_is_included(self):
-        self.assertEqual(
-            self.returned_data[6][self.OUTPUT_COLUMN],
-            self.expected_data[6][self.OUTPUT_COLUMN],
-        )
-
-    def test_returned_df_matches_expected_df(self):
-        self.assertEqual(
-            self.returned_data,
-            self.expected_data,
-        )
-
-    def test_returned_df_has_one_additional_column(self):
-        self.assertEqual(
-            len(self.returned_df.columns), len(self.test_purge_outdated_df.columns) + 1
-        )
-
-    def test_returned_df_has_same_number_of_rows(self):
-        self.assertEqual(self.returned_df.count(), self.test_purge_outdated_df.count())
-
-
 class RemoveLocationsWithDuplicatesTests(IngestASCWDSWorkerDatasetTests):
     def setUp(self) -> None:
         super().setUp()
